automobile_django/requirements.txt
Django==1.11.2
django-braces==1.11.0
django-cors-headers==2.1.0
django-ordered-model==1.4.1
graphene==1.4
graphene-django==1.3
graphql-core==1.1
graphql-relay==0.4.5
iso8601==0.1.11
promise==2.0.2
pytz==2017.2
singledispatch==3.4.0.3
six==1.10.0
typing==3.6.1
